+++ public/data/data.json
@@ -3,6 +3,7 @@
     "id": 1,
     "name": "Leanne Graham",
     "username": "Bret",
+    "imgUrl": "https://images.unsplash.com/photo-1667833183058-c13008862b87?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=880&q=80",
     "email": "Sincere@april.biz",
     "address": {
       "street": "Kulas Light",
@@ -26,6 +27,7 @@
     "id": 2,
     "name": "Ervin Howell",
     "username": "Antonette",
+    "imgUrl": "https://images.unsplash.com/photo-1668776157628-e26cbeed6eb4?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=808&q=80",
     "email": "Shanna@melissa.tv",
     "address": {
       "street": "Victor Plains",
@@ -49,6 +51,7 @@
     "id": 3,
     "name": "Clementine Bauch",
     "username": "Samantha",
+    "imgUrl": "https://images.unsplash.com/photo-1664994817539-8a865f12dc6c?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=765&q=80",
     "email": "Nathan@yesenia.net",
     "address": {
       "street": "Douglas Extension",
@@ -72,6 +75,7 @@
     "id": 4,
     "name": "Patricia Lebsack",
     "username": "Karianne",
+    "imgUrl": "https://images.unsplash.com/photo-1668530933925-a57e8424cda0?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=764&q=80",
     "email": "Julianne.OConner@kory.org",
     "address": {
       "street": "Hoeger Mall",
@@ -95,6 +99,7 @@
     "id": 5,
     "name": "Chelsey Dietrich",
     "username": "Kamren",
+    "imgUrl": "https://images.unsplash.com/photo-1668670579277-4b3ffa47b016?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=788&q=80",
     "email": "Lucio_Hettinger@annie.ca",
     "address": {
       "street": "Skiles Walks",
@@ -118,6 +123,7 @@
     "id": 6,
     "name": "Mrs. Dennis Schulist",
     "username": "Leopoldo_Corkery",
+    "imgUrl": "https://images.unsplash.com/photo-1669725337418-f81459093abd?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=880&q=80",
     "email": "Karley_Dach@jasper.info",
     "address": {
       "street": "Norberto Crossing",
@@ -141,6 +147,7 @@
     "id": 7,
     "name": "Kurtis Weissnat",
     "username": "Elwyn.Skiles",
+    "imgUrl": "https://images.unsplash.com/photo-1669406780538-e57f8c0b248e?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=880&q=80",
     "email": "Telly.Hoeger@billy.biz",
     "address": {
       "street": "Rex Trail",
@@ -164,6 +171,7 @@
     "id": 8,
     "name": "Nicholas Runolfsdottir V",
     "username": "Maxime_Nienow",
+    "imgUrl": "https://images.unsplash.com/photo-1669467566337-199d90432686?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=764&q=80",
     "email": "Sherwood@rosamond.me",
     "address": {
       "street": "Ellsworth Summit",
@@ -187,6 +195,7 @@
     "id": 9,
     "name": "Glenna Reichert",
     "username": "Delphine",
+    "imgUrl": "https://images.unsplash.com/photo-1667736697093-b42d29d0dc73?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=880&q=80",
     "email": "Chaim_McDermott@dana.io",
     "address": {
       "street": "Dayna Park",
@@ -210,6 +219,7 @@
     "id": 10,
     "name": "Clementina DuBuque",
     "username": "Moriah.Stanton",
+    "imgUrl": "https://images.unsplash.com/photo-1667835949430-a2516cc93d27?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=764&q=80",
     "email": "Rey.Padberg@karina.biz",
     "address": {
       "street": "Kattie Turnpike",
