+++ src/pages/kimnahee/Main/Main.scss
@@ -72,7 +72,8 @@ $MainPrimary-border: 2px solid #efefef;
 .main {
   @include flexSort(center);
   width: 100%;
-  height: 110vh;
+  height: auto;
+  padding-bottom: 50px;
   background-color: rgb(240, 240, 240);
 }
